Add a separate thread to handle read requests
Read requests are now sent through a SelectRequest struct that couples
the select statement together with a channel to return the
results.

Additionally, the index is now wrapped in a mutex that's itself
wrapped in a thread-safe reference-counting pointer. This enables the
index to be read/written in multiple threads. There are performance
tradeoffs to this because the entire index has to be locked on
read/write, so one possibility is that we could shard the index into
multiple hashmaps.