Replace mutex over index with a reader-writer lock
A reader-writer lock allows us to hold multiple read locks on the
index, which is the usual use case when inserting data into existing
time series, or when reading data from existing time series. Write
locks on the index are only held when a completely new series is being
added. This solves the performance issue of the entire index being
locked whenever a read/write occurs.