Implement multi-threaded read and writes (#3)
* Add Clone traits to select structs

* Make the Select struct public

* Add a separate thread to handle read requests

Read requests are now sent through a SelectRequest struct that couples
the select statement together with a channel to return the
results.

Additionally, the index is now wrapped in a mutex that's itself
wrapped in a thread-safe reference-counting pointer. This enables the
index to be read/written in multiple threads. There are performance
tradeoffs to this because the entire index has to be locked on
read/write, so one possibility is that we could shard the index into
multiple hashmaps.

* Replace mutex over index with a reader-writer lock

A reader-writer lock allows us to hold multiple read locks on the
index, which is the usual use case when inserting data into existing
time series, or when reading data from existing time series. Write
locks on the index are only held when a completely new series is being
added. This solves the performance issue of the entire index being
locked whenever a read/write occurs.